feat(dapp): Refactor auction/lib/utils
