+++ dapp/src/auctions/hooks/useAuctionBid.ts
@@ -7,7 +7,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useIotaNamesClient } from '@/contexts';
 import { queryKey } from '@/hooks/queryKey';
 
-import { buildCreateAuctionTransaction, buildPlaceBidTransaction } from '../lib/transaction';
+import { buildCreateAuctionTransaction, buildPlaceBidTransaction } from '../lib/utils/transaction';
 import { useAuctionHouse } from './useAuctionHouse';
 import { useGetAuctionMetadata } from './useGetAuctionMetadata';
