+++ dapp/src/auctions/hooks/useClaimAuctionTransaction.ts
@@ -8,7 +8,7 @@ import { useIotaNamesClient } from '@/contexts';
 import { queryKey } from '@/hooks/queryKey';
 import { getGasSummary } from '@/lib/utils/getGasSummary';
 
-import { buildClaimDomainTransaction } from '../lib/transaction';
+import { buildClaimDomainTransaction } from '../lib/utils/transaction';
 import { useAuctionHouse } from './useAuctionHouse';
 
 export function useClaimAuctionTransaction(address: string, domain: string) {
