+++ dapp/src/auctions/lib/types/bcs.ts
