+++ dapp/src/auctions/lib/types/index.ts
@@ -1,8 +1,5 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './transaction';
 export * from './bcs';
-export * from './utils';
 export * from './metadata';
-export * from './types';
