+++ dapp/src/auctions/lib/utils/metadata.ts
@@ -4,11 +4,12 @@
 import { getDomainType } from '@iota/iota-names-sdk';
 import { IotaGraphQLClient } from '@iota/iota-sdk/graphql';
 import { graphql } from '@iota/iota-sdk/graphql/schemas/2025.2';
-import { fromB64 } from '@iota/iota-sdk/utils';
+import { fromB64, NANOS_PER_IOTA } from '@iota/iota-sdk/utils';
 
-import { AuctionFieldBcs } from './bcs';
-import { AuctionMetadata, normalizeAuctionData } from './types';
-import { createDomainFromName, deriveAuctionDynamicFieldId } from './utils';
+import { AuctionFieldBcs } from '../types/bcs';
+import { AuctionMetadata, RawAuctionMetadata } from '../types/metadata';
+import { deriveAuctionDynamicFieldId } from './dynamicField';
+import { createDomainFromName } from './name';
 
 interface FetchAuctionMetadataParams {
     domainName: string;
@@ -67,9 +68,34 @@ export async function fetchAuctionMetadata({
 }
 
 /**
+ * Converts raw BCS auction data to normalized TypeScript interface
+ */
+function normalizeAuctionData(rawData: RawAuctionMetadata): AuctionMetadata | null {
+    if (!rawData?.value?.value) {
+        return null;
+    }
+
+    const auction = rawData.value.value;
+
+    const currentBidNanos = BigInt(auction.current_bid.balance.value || BigInt(0));
+    const minBidNanos = currentBidNanos + NANOS_PER_IOTA;
+
+    return {
+        domainName: auction.nft.domain_name,
+        startTimestamp: new Date(Number(auction.start_timestamp_ms)),
+        endTimestamp: new Date(Number(auction.end_timestamp_ms)),
+        winner: auction.winner,
+        currentBidNanos,
+        minBidNanos,
+        nftId: auction.nft.id,
+        nftExpiration: new Date(Number(auction.nft.expiration_timestamp_ms)),
+    };
+}
+
+/**
  * Parses raw auction data into the bcs type.
  */
-export function parseAuctionDataBcs(
+function parseAuctionDataBcs(
     data: Awaited<ReturnType<typeof fetchAuctionMetadata>>,
 ): ReturnType<typeof AuctionFieldBcs.parse> | null {
     const auctionBcsB64 = data?.object?.asMoveObject?.contents?.bcs;
